MythxJS added isAccessTokenValid check
